fix: removing pointless cast

+++ hdr/json.h
@@ -11,7 +11,7 @@
 void save_queue(const std::queue<std::string>& q, const std::string& filename);
 std::queue<std::string> load_queue(const std::string& filename);
 
-uint16_t get_offset(const std::string& filename);
+uint64_t get_offset(const std::string& filename);
 void increment_offset(const std::string& filename);
 
 std::string get_token(const std::string& filename);
